admin/requirements.txt
django==4.1.2
djangorestframework==3.14.0
pika==1.1.0
django-cors-headers==3.5.0
psycopg2>=2.8
drf_yasg